renamed DataspaceRegistrationPolicy
